Implement task graph caching (TGC)
Task relationships are still represented without repetition in the
canonical entity data (as :subtask-ids), but the entity cache will now
store a precalculated record of each task's parent task as well. In
short, we have bidirectionally linked performance with
single-direction linking API simplicity.

As a first application of this cache, we will now show the ancestry of
each task in its task list item. These should probably be ellipsized
but are not now.

This is also the first commit which makes the actual assumption that
each task has only a single parent. As long as subtasking refers only
to composition, and does not need to be the sole way to draw
cross-task dependencies, I think this is fine.

Finally, the API would make more sense if we used :parent-task-id
instead of :subtask-ids, since the former naturally prevents multiple
parents.