Remove context-stack representation in code
Since we can now reliably and quickly poll a task's ancestry chain, we
can just store the 'selected' task and no longer need to keep a stack
or breadcrumb to where we came from. This is nice because it
simplifies the code & actually fixes an issue where, when subtasks
were selected from the 'milestones' menu, they would be missing their
parent context items.