Fix nREPL access through Docker
It was a nightmare getting nREPL to work between CIDER on host Emacs
and nREPL in container. Now, fighwheel-main no longer works. It won't
start serving at init (because main was swapped to nREPL) and for some
reason it won't launch manually from the nREPL client either. Fixing
that next.